requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
jsondiff==2.0.0
multidict==6.0.5
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2023.4
requests==2.31.0
six==1.16.0
TgCrypto==1.2.5
tornado==6.4
tzlocal==5.2
urllib3==2.2.0
yarl==1.9.4
